moved no config into correct package
