+++ src/main/java/com/github/onsdigital/logging/v2/config/Config.java
@@ -2,7 +2,6 @@ package com.github.onsdigital.logging.v2.config;
 
 import com.github.onsdigital.logging.v2.Logger;
 import com.github.onsdigital.logging.v2.LoggingException;
-import com.github.onsdigital.logging.v2.config.nop.LogConfig;
 import com.github.onsdigital.logging.v2.serializer.LogSerialiser;
 import com.github.onsdigital.logging.v2.storage.LogStore;
 import org.apache.commons.lang3.StringUtils;
