+++ src/main/java/com/github/onsdigital/logging/v2/config/nop/NopConfig.java
@@ -1,6 +1,7 @@
 package com.github.onsdigital.logging.v2.config.nop;
 
 import com.github.onsdigital.logging.v2.Logger;
+import com.github.onsdigital.logging.v2.config.LogConfig;
 import com.github.onsdigital.logging.v2.config.ShutdownHook;
 import com.github.onsdigital.logging.v2.serializer.LogSerialiser;
 import com.github.onsdigital.logging.v2.storage.LogStore;
